requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
asyncio
requests==2.31.0
aiohttp==3.9.1
